transfer debug console to proper file
